requirements.txt
httpx==0.27.0
Pillow==10.3.0
python-dotenv==1.0.1
streamlit>=1.35.0
markdown>=3.6